+++ layout/columns2.php
@@ -101,5 +101,8 @@ require_once(__DIR__ . '/includes/staticpages.php');
 // Include the template content for the JavaScript disabled hint.
 require_once(__DIR__ . '/includes/javascriptdisabledhint.php');
 
+// Include the template content for the info banners.
+require_once(__DIR__ . '/includes/infobanners.php');
+
 // Render columns2.mustache from boost_union.
 echo $OUTPUT->render_from_template('theme_boost_union/columns2', $templatecontext);