+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2022-07-15 - Feature: Information banners (base functionality), helps to resolve #4.
 * 2022-07-15 - Settings: Divide theme settings into multiple pages, solves #52
 * 2022-07-13 - Release: Add contributors to README.md, solves #44
 * 2022-07-13 - Release: Add UPGRADE.md, solves #29